feat (invitation): Añadida gestion de invitaciones
Añadida gestion de invitaciones a falta de validaciones e inclusión de "spring.jackson.serialization.fail-on-empty-beans=false" en application.properties para solucionar problema con hibernate lazy initializer (al pasar objeto user y team a dto)

+++ src/main/java/com/spring/Repository/TeamRepository.java
@@ -2,9 +2,8 @@ package com.spring.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.spring.CustomObject.TeamDto;
 import com.spring.Model.Team;
 
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 
 }
